Restructure + question section
